- major changes in code to make it work with the EPMT package system. - the earlier code relied on running an executable in the dash directory - the new code uses relative packages to be imported and executed.
The recipe is to be found in epmt_cmds, but in short it is:

>>>     from ui.index import init_app
>>>     init_app()
>>>     from ui.app import app
>>>     app.run_server(debug=False, host='0.0.0.0')

This has to be executed from the top-level EPMT directory.

Also fixed bug in detecting matching models. The earlier code unnecessarily
relied on dumps of a dict and was getting confused with out of order dict
keys.

The new code now works for SQLA and Pony.